// src/services/api/models/buy-order.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { OrderType } from './order-type';
import { SaleType } from './sale-type';
/**
 * 
 * @export
 * @interface BuyOrder
 */
export interface BuyOrder {
    /**
     * 
     * @type {string}
     * @memberof BuyOrder
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof BuyOrder
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof BuyOrder
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof BuyOrder
     */
    updated?: Date;
    /**
     * 
     * @type {string}
     * @memberof BuyOrder
     */
    saleId?: string;
    /**
     * 
     * @type {string}
     * @memberof BuyOrder
     */
    saleParentId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof BuyOrder
     */
    userId?: string;
    /**
     * 
     * @type {SaleType}
     * @memberof BuyOrder
     */
    saleType?: SaleType;
    /**
     * 
     * @type {OrderType}
     * @memberof BuyOrder
     */
    orderType?: OrderType;
    /**
     * 
     * @type {number}
     * @memberof BuyOrder
     */
    amount?: number;
}