// src/services/api/models/sale-asset.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Blockchain } from './blockchain';
import { RedemptionStatus } from './redemption-status';
/**
 * 
 * @export
 * @interface SaleAsset
 */
export interface SaleAsset {
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof SaleAsset
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof SaleAsset
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof SaleAsset
     */
    updated?: Date;
    /**
     * 
     * @type {Blockchain}
     * @memberof SaleAsset
     */
    blockchain?: Blockchain;
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    ownerAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    contractAddress?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    tokenId?: string | null;
    /**
     * 
     * @type {RedemptionStatus}
     * @memberof SaleAsset
     */
    redemptionStatus?: RedemptionStatus;
    /**
     * 
     * @type {boolean}
     * @memberof SaleAsset
     */
    isTransferable?: boolean;
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    saleId?: string;
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    walletAssetId?: string;
    /**
     * 
     * @type {string}
     * @memberof SaleAsset
     */
    assetPackId?: string;
}