// src/services/api/models/update-auction-end-time-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpdateAuctionEndTimeRequest
 */
export interface UpdateAuctionEndTimeRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateAuctionEndTimeRequest
     */
    saleId?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAuctionEndTimeRequest
     */
    liveStreamId?: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateAuctionEndTimeRequest
     */
    buyOrderId?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof UpdateAuctionEndTimeRequest
     */
    auctionEndTime?: Date;
    /**
     * 
     * @type {string}
     * @memberof UpdateAuctionEndTimeRequest
     */
    assetName?: string | null;
}