// src/services/api/models/update-auction-start-time-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface UpdateAuctionStartTimeRequest
 */
export interface UpdateAuctionStartTimeRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateAuctionStartTimeRequest
     */
    saleId?: string;
    /**
     * 
     * @type {number}
     * @memberof UpdateAuctionStartTimeRequest
     */
    auctionEndSeconds?: number;
}